+++ seedfarmer/__main__.py
@@ -23,35 +23,10 @@ import yaml
 import seedfarmer.mgmt.deploy_utils as du
 import seedfarmer.mgmt.module_info as mi
 import seedfarmer.mgmt.module_init as minit
-from seedfarmer import DESCRIPTION, PROJECT, commands, utils
+from seedfarmer import DEBUG_LOGGING_FORMAT, DESCRIPTION, PROJECT, commands, enable_debug, utils
 from seedfarmer.output_utils import print_bolded, print_deployment_inventory, print_json, print_manifest_inventory
 
-DEBUG_LOGGING_FORMAT = "[%(asctime)s][%(filename)-13s:%(lineno)3d] %(message)s"
-DEBUG_LOGGING_FORMAT_REMOTE = "[%(filename)-13s:%(lineno)3d]%(message)s"
-
-INFO_LOGGING_FORMAT = "[%(filename)-13s:%(lineno)3d] %(message)s"
-INFO_LOGGING_FORMAT_REMOTE = "[%(filename)-13s:%(lineno)3d]%(message)s"
 _logger: logging.Logger = logging.getLogger(__name__)
-
-
-def enable_debug(format: str) -> None:
-    logging.basicConfig(level=logging.DEBUG, format=format)
-    _logger.setLevel(logging.DEBUG)
-    logging.getLogger("boto3").setLevel(logging.ERROR)
-    logging.getLogger("botocore").setLevel(logging.ERROR)
-    logging.getLogger("s3transfer").setLevel(logging.ERROR)
-    logging.getLogger("urllib3").setLevel(logging.ERROR)
-    logging.getLogger("sh").setLevel(logging.ERROR)
-
-
-def enable_info(format: str) -> None:
-    logging.basicConfig(level=logging.INFO, format=format)
-    _logger.setLevel(logging.INFO)
-    logging.getLogger("boto3").setLevel(logging.ERROR)
-    logging.getLogger("botocore").setLevel(logging.ERROR)
-    logging.getLogger("s3transfer").setLevel(logging.ERROR)
-    logging.getLogger("urllib3").setLevel(logging.ERROR)
-    logging.getLogger("sh").setLevel(logging.ERROR)
 
 
 @click.group()
@@ -60,7 +35,7 @@ def cli() -> None:
     pass
 
 
-@click.command(help=f"Apply a deployment spec relative path for {PROJECT.upper()}")
+@click.command(help=f"Apply a deployment manifest relative path for {PROJECT.upper()}")
 @click.argument(
     "spec",
     type=str,
@@ -86,12 +61,9 @@ def cli() -> None:
 )
 def apply(spec: str, debug: bool, dry_run: bool, show_manifest: bool) -> None:
     f"Deploy a(n) {PROJECT.upper()} environemnt based on a deployspec file (yaml)." ""
-
     if debug:
         enable_debug(format=DEBUG_LOGGING_FORMAT)
-    else:
-        enable_info(format=INFO_LOGGING_FORMAT)
-    _logger.info(f"Apply request with spec {spec} ")
+    _logger.info(f"Apply request with manifest {spec} ")
     if dry_run:
         print_bolded(" ***   This is a dry-run...NO ACTIONS WILL BE TAKEN  *** ", "white")
 
@@ -131,8 +103,6 @@ def destroy(
     f"Destroy an {PROJECT.upper} deployment ."
     if debug:
         enable_debug(format=DEBUG_LOGGING_FORMAT)
-    else:
-        enable_info(format=INFO_LOGGING_FORMAT)
     _logger.info(f"Destroy for {deployment} ")
     if dry_run:
         print_bolded(" ***   This is a dry-run...NO ACTIONS WILL BE TAKEN  *** ", "white")
@@ -396,9 +366,19 @@ def list_module_metadata(
     help="The Deployment Name",
     required=True,
 )
+@click.option(
+    "--debug/--no-debug",
+    default=False,
+    help="Enable detailed logging.",
+    show_default=True,
+)
 def list_modules(
     deployment: str,
+    debug: bool,
 ) -> None:
+    if debug:
+        enable_debug(format=DEBUG_LOGGING_FORMAT)
+    _logger.debug(f"We are getting modules for  {deployment}")
 
     cache = du.generate_deployment_cache(deployment_name=deployment)
     dep_manifest = du.generate_deployed_manifest(
@@ -478,39 +458,23 @@ def init() -> None:
     pass
 
 
-@init.command(name="project", help="Initialize a project")
-@click.option(
-    "--project-name",
-    "-n",
-    type=str,
-    help="The name of the project",
-    required=True,
-    default=None,
+@init.command(
+    name="project",
+    help=(
+        "Initialize a project. " "Make sure seedfarmer.yaml is present in the same location you execute this command!!"
+    ),
 )
 @click.option(
     "--template-url",
     "-t",
-    default="git@ssh.gitlab.aws.dev:wwcs-proserve-etip-data-analytics/software-labs/seed-farmer.git",
+    default="git@github.com:awslabs/seed-farmer.git",
     help=(
-        "The template URL. If not specified, the default template repo is "
-        "`https://gitlab.aws.dev/wwcs-proserve-etip-data-analytics/software-labs/seed-farmer`"
+        "The template URL. If not specified, the default template repo is " "`https://github.com/awslabs/seed-farmer`"
     ),
     required=False,
 )
-@click.option(
-    "--no-interactive-input/--interactive-input",
-    default=True,
-    help="Enable interactive prompt at command execution. Default is False",
-    required=False,
-)
-@click.option("--debug/--no-debug", default=False, help="Enable detail logging", show_default=True)
-def init_project(project_name: str, template_url: str, no_interactive_input: bool, debug: bool) -> None:
-    if debug:
-        enable_debug(format=DEBUG_LOGGING_FORMAT)
-    _logger.debug(f"Initializing project {project_name}")
-    minit.create_project(
-        project_name=project_name, template_url=template_url, no_interactive_input=no_interactive_input
-    )
+def init_project(template_url: str) -> None:
+    minit.create_project(template_url=template_url)
 
 
 @init.command(name="module", help="Initialize a new module")
@@ -526,21 +490,14 @@ def init_project(project_name: str, template_url: str, no_interactive_input: boo
 @click.option(
     "--template-url",
     "-t",
-    default="git@ssh.gitlab.aws.dev:wwcs-proserve-etip-data-analytics/software-labs/seed-farmer.git",
+    default="git@github.com:awslabs/seed-farmer.git",
     help=(
-        "The template URL. If not specified, the default template repo is "
-        "`https://gitlab.aws.dev/wwcs-proserve-etip-data-analytics/software-labs/seed-farmer`"
+        "The template URL. If not specified, the default template repo is " "`https://github.com/awslabs/seed-farmer`"
     ),
     required=False,
 )
-@click.option(
-    "--no-interactive-input/--interactive-input",
-    default=True,
-    help="Enable interactive prompt at command execution. Default is False",
-    required=False,
-)
 @click.option("--debug/--no-debug", default=False, help="Enable detail logging", show_default=True)
-def init_module(group_name: str, module_name: str, template_url: str, no_interactive_input: bool, debug: bool) -> None:
+def init_module(group_name: str, module_name: str, template_url: str, debug: bool) -> None:
     if debug:
         enable_debug(format=DEBUG_LOGGING_FORMAT)
     _logger.debug(f"Initializing module {module_name}")
@@ -548,7 +505,6 @@ def init_module(group_name: str, module_name: str, template_url: str, no_interac
         group_name=group_name,
         module_name=module_name,
         template_url=template_url,
-        no_interactive_input=no_interactive_input,
     )