+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 
 ---
-## 0.1.1.dev0 -- Unreleased
+## v0.1.1 (2022-07-07)
 
 ### New
  - feature - added CLI ability to list deployspec of deployed module
@@ -17,11 +17,19 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### Changes
 - change - enforce runtime versions of CodeSeeder
+- lock versions in requirements-dev.in
+- updating CLI output to discern changes to manifest, intended deployment
+- lazy-load seedkit (version 0.3.2 of codeseeder)
+- improve logging messages from codeseeder 
+- align message output / tables listing modules requested, modified, deleted for deployment
 
 ### Fixes
+- update repo for project initialization
+- update repo for module initalizaton
+- optimize intialization (remove interactive input, require seedfarmer.yaml)
 
 ---
-## 0.1.0 -- (2022-06-20)
+## v0.1.0  (2022-06-20)
 
 
 ### New